text_summary/pyproject.toml
[tool.poetry]
name = "text_summary"
version = "0.1.0"
description = ""
authors = ["aevtikheev <eft000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.63.0"
uvicorn = "0.13.4"
asyncpg = "0.22.0"
tortoise-orm = "0.16.21"
gunicorn = "20.0.4"
newspaper3k = "^0.2.8"
aerich = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "6.2.2"
requests = "2.25.1"
pytest-cov = "2.10.1"
flake8 = "3.8.4"
pytest-xdist = "^2.2.1"
flake8-2020 = "^1.6.0"
flake8-broken-line = "^0.3.0"
flake8-bugbear = "^21.3.2"
flake8-commas = "^2.0.0"
flake8-comprehensions = "^3.4.0"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.0.0"
flake8-fixme = "^1.1.1"
flake8-multiline-containers = "^0.0.17"
flake8-print = "^4.0.0"
flake8-return = "^1.1.2"
flake8-quotes = "^3.2.0"
flake8-simplify = "^0.14.0"
flake8-variables-names = "^0.0.4"
flake8-walrus = "^1.1.0"
pep8-naming = "^0.11.1"
flake8-cognitive-complexity = "^0.1.0"
flake8-expression-complexity = "^0.0.9"
flake8-docstrings = "^1.5.0"
flake8-markdown = "^0.2.0"
flake8-tidy-imports = "^4.2.1"
flake8-import-order = "^0.18.1"
flake8-bandit = "^2.1.2"
flake8-annotations-complexity = "^0.0.6"
flake8-annotations-coverage = "^0.0.5"
flake8-string-format = "^0.3.0"
mccabe = "^0.6.1"
mypy = "^0.812"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
